Fix compilation warning

+++ utils/fasm/test/test_fasm.cpp
@@ -33,12 +33,12 @@ TEST_CASE("fasm_integration_test", "[fasm]") {
         REQUIRE(flow_succeeded == true);
 
         auto &device_ctx = g_vpr_ctx.mutable_device();
-        for(int inode = 0; inode < device_ctx.rr_nodes.size(); ++inode) {
+        for(size_t inode = 0; inode < device_ctx.rr_nodes.size(); ++inode) {
             for(int iedge = 0; iedge < device_ctx.rr_nodes[inode].num_edges(); ++iedge) {
                 auto sink_inode = device_ctx.rr_nodes[inode].edge_sink_node(iedge);
                 auto switch_id = device_ctx.rr_nodes[inode].edge_switch(iedge);
                 vpr::add_rr_edge_metadata(inode, sink_inode, switch_id,
-                        "fasm_features", vtr::string_fmt("%d_%d_%d",
+                        "fasm_features", vtr::string_fmt("%d_%d_%zu",
                             inode, sink_inode, switch_id));
             }
         }
